1st draft of complete bug fix summary of axionCAMB
Including the specific bug fix lines and comparison of the observables before and after debugging. Designed to be a full instruction that anyone with a published axionCAMB can reproduce, this notebook also includes a quick 5-min copy-and-past subprocess implementation enabling calling axionCAMB from the jupyter notebook collectively with a specified range of parameters of interest.